Clean up namespaces for dead trees elimination
Since optimization classes use dynamic inheritance with virtual methods
by default, there is no need to create an optimization class under OMR
namespace and then map to TR namespace.
