+++ compiler/optimizer/DeadTreesElimination.cpp
