+++ compiler/optimizer/DeadTreesElimination.hpp
@@ -19,19 +19,82 @@
 #ifndef TR_DEADTREESELIMINATION_INCL
 #define TR_DEADTREESELIMINATION_INCL
 
-#include "optimizer/OMRDeadTreesElimination.hpp"
+#include <stdint.h>                    // for int32_t, uint32_t
+#include "env/TRMemory.hpp"            // for Allocator, TR_Memory, etc
+#include "infra/List.hpp"              // for TR_ScratchList
+#include "optimizer/Optimization.hpp"  // for Optimization
 
+namespace TR { class Block; }
 namespace TR { class OptimizationManager; }
+namespace TR { class TreeTop; }
+
+namespace OMR
+{
+
+class TreeInfo
+   {
+   public:
+   TR_ALLOC(TR_Memory::LocalOpts)
+
+   TreeInfo(TR::TreeTop *treeTop, int32_t height)
+      : _tree(treeTop),
+        _height(height)
+      {
+      }
+
+   TR::TreeTop *getTreeTop()  {return _tree;}
+   void setTreeTop(TR::TreeTop *tree)  {_tree = tree;}
+
+   int32_t getHeight()    {return _height;}
+   void setHeight(int32_t height) {_height = height;}
+
+   private:
+
+   int32_t _height;
+   TR::TreeTop *_tree;
+   };
+
+}
+
 
 namespace TR
 {
 
-class DeadTreesElimination : public OMR::DeadTreesElimination
+/*
+ * Class DeadTreesElimination
+ * ==========================
+ *
+ * Expressions that are evaluated and not used for a few instructions can 
+ * actually be evaluated exactly where required (subject to the constraint 
+ * that their values should be the same if evaluated later). This optimization 
+ * attempts to simply delay evaluation of expressions and has the effect of 
+ * reducing register pressure by shortening live ranges. It has the effect 
+ * of removing some IL trees that are simply anchors for (already) evaluated 
+ * expressions and making the trees more compact.
+ */
+
+class DeadTreesElimination : public TR::Optimization
    {
    public:
 
-   DeadTreesElimination(TR::OptimizationManager *manager) :
-      OMR::DeadTreesElimination(manager) {}
+   DeadTreesElimination(TR::OptimizationManager *manager);
+   static TR::Optimization *create(TR::OptimizationManager *manager);
+
+   virtual int32_t perform();
+   virtual int32_t performOnBlock(TR::Block *);
+   virtual void prePerformOnBlocks();
+
+   protected:
+
+   virtual TR::TreeTop *findLastTreetop(TR::Block *block, TR::TreeTop *prevTree);
+
+   private:
+
+   int32_t process(TR::TreeTop *, TR::TreeTop *);
+
+   List<OMR::TreeInfo> _targetTrees;
+   bool _cannotBeEliminated;
+   bool _delayedRegStores;
    };
 
 }
